Added schema generator and fixed various minor errors. - Fixed quoted HLP parser not ingesting last character. - Added types conversions to HLP from ECS types. - Fixed bad_any_cast of HLP operation when converting float to double. - Added python script to generate schemas and various files. - Included AWS ELB ECS fields to the script and generated schemas. - Added posibility to use definitions inside definitions.

+++ src/engine/source/builder/builders/opBuilderLogqlParser.cpp
@@ -37,7 +37,7 @@ static bool any2Json(std::any const& anyVal, std::string const& path, base::Even
     else if (typeid(float) == type || typeid(double) == type)
     {
         json::Json val;
-        val.setDouble(std::any_cast<double>(anyVal));
+        val.setDouble(std::any_cast<float>(anyVal));
         event->set(path, val);
     }
     else if (typeid(std::string) == type)
